requirements.txt
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
pandas==1.3.3
openpyxl==3.0.9
yfinance==0.1.63
finnhub-python==1.2.2
